websockets/interfaces/WebSocket/send/003.html

<!doctype html>
<title>WebSockets: send() when readyState is CONNECTING</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
  assert_throws("INVALID_STATE_ERR", function(){ws.send('a')});
});
</script>

